all passing tests

+++ migrations/sqls/20210912212103-products-table-up.sql
@@ -1,2 +1,2 @@
 /* Replace with your SQL commands */
-CREATE TABLE products(id SERIAL PRIMARY KEY, name VARCHAR, price integer);
+CREATE TABLE products(id SERIAL PRIMARY KEY, name VARCHAR, price decimal);
